Adding back an older release and making it somewhat conservative
I've decided ReBOOM with all the bells I whistles I added was just too unstable.  This is a version right from 2.02, ported to SDL2 with the always hud stat on and the fullscreen toggle switch from Choco.  Consider this "a conservative port".  It is a replacement for BOOM.EXE on DOS.  Not that anyone will take notice because they are all too arrogant and elitist.